@@ -7,7 +7,7 @@ readme = "README.md"
 packages = [{include = "src"}]
 
 [tool.poetry.dependencies]
-python = "^3.9"
+python = "^3.13"
 google-auth = "^2.23.0"
 google-auth-oauthlib = "^1.1.0"
 google-auth-httplib2 = "^0.1.1"
@@ -18,8 +18,7 @@ requests = "^2.31.0"
 [tool.poetry.group.dev.dependencies]
 pytest = "^7.4.0"
 pytest-cov = "^4.1.0"
-black = "^23.7.0"
-flake8 = "^6.0.0"
+ruff = "^0.6.0"
 mypy = "^1.5.0"
 
 [build-system]
@@ -47,3 +46,44 @@ ignore_missing_imports = true
 [[tool.mypy.overrides]]
 module = "tests.*"
 ignore_errors = true
+
+[tool.ruff]
+# Same as Black's default line length
+line-length = 88
+target-version = "py39"
+
+# Exclude directories
+exclude = [
+    ".git",
+    "__pycache__",
+    ".aws-sam",
+    "package",
+]
+
+[tool.ruff.lint]
+# Enable specific rule categories
+select = [
+    "E",   # pycodestyle errors
+    "W",   # pycodestyle warnings
+    "F",   # Pyflakes
+    "I",   # isort
+    "N",   # pep8-naming
+    "UP",  # pyupgrade
+    "B",   # flake8-bugbear
+    "C4",  # flake8-comprehensions
+    "SIM", # flake8-simplify
+]
+
+# Ignore specific rules that conflict with Black or are too strict
+ignore = [
+    "E203", # whitespace before ':'
+]
+
+[tool.ruff.lint.isort]
+known-first-party = ["src", "shared"]
+
+[tool.ruff.format]
+# Use single quotes instead of double quotes
+quote-style = "double"
+# Respect magic trailing commas
+skip-magic-trailing-comma = false
